+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 ### Unreleased
 
+## v2.1.0 (2021-03-08)
+
 * Require that the config values for session hash salt, garbage collect token and tokenista secret
   are defined (not null) before the dependency container creates the related service.
 * Add extensible HTTP-triggered garbage collection (see README)